feat(design-system): extract palette generator → @klorad/design-system/palette
PR 2 of the Mobility arc. Same pattern as PR 1 (@klorad/secrets):
move a domain-agnostic helper from apps/campus/lib/ into a shared
package so every Klorad vertical reuses it without forking.

  • New sub-export @klorad/design-system/palette
    derivePalette(hex) → BrandPalette
    paletteToCssVars(palette) → CSSProperties
    Generalised from the Campus-named originals; same OKLCH math,
    same eight tokens (primary + fill/bg/soft/ink + warm/cool/
    complement), same --brand-* CSS variable naming convention.

  • Sub-exported (not in the main entry) so consumers that don't
    theme keep culori out of their bundle. culori dep migrates to
    @klorad/design-system; dropped from apps/campus alongside
    @types/culori (was only used by lib/palette).

  • apps/campus/lib/palette.ts becomes a re-export shim that
    aliases the generalised names back to deriveCampusPalette /
    CampusPalette so the two existing import sites (consumer
    layout + identity settings) keep working without an edit.

Zero behavioural change for Campus. Mobility will import
@klorad/design-system/palette directly when its identity screen
lands.
