+++ packages/design-system/package.json
@@ -7,11 +7,13 @@
   "types": "./src/index.ts",
   "exports": {
     ".": "./src/index.ts",
+    "./palette": "./src/palette/index.ts",
     "./tailwind-preset": "./src/tailwind-preset.ts",
     "./tokens.css": "./src/styles/tokens.css"
   },
   "dependencies": {
     "clsx": "^2.1.1",
+    "culori": "^4.0.2",
     "next-themes": "^0.4.6",
     "tailwind-merge": "^3.6.0"
   },
